+++ src/okta/createOktaClient.ts
@@ -7,18 +7,82 @@ import { OktaIntegrationConfig } from '../types';
 const okta = require('@okta/okta-sdk-nodejs');
 const MemoryStore = require('@okta/okta-sdk-nodejs/src/memory-store');
 
-// Keep retrying until the Lambda times out
-const OKTA_MAX_RETRIES = 0;
-const OKTA_REQUEST_TIMEOUT = 0;
+const DEFAULT_MINIMUM_RATE_LIMIT_REMAINING = 5;
+
+interface RequestExecutorWithEarlyRateLimitingOptions {
+  minimumRateLimitRemaining?: number;
+}
+
+/**
+ * A custom Okta request executor that throttles requests when `x-rate-limit-remaining` response
+ * headers fall below a provided threshold.
+ */
+export class RequestExecutorWithEarlyRateLimiting extends okta.RequestExecutor {
+  logger: IntegrationLogger;
+  minimumRateLimitRemaining: number;
+  requestAfter: number | undefined;
+
+  constructor(
+    logger: IntegrationLogger,
+    options?: RequestExecutorWithEarlyRateLimitingOptions,
+  ) {
+    super();
+    this.logger = logger;
+    this.minimumRateLimitRemaining =
+      options?.minimumRateLimitRemaining ||
+      DEFAULT_MINIMUM_RATE_LIMIT_REMAINING;
+  }
+
+  async fetch(request: any) {
+    const now = Date.now();
+    if (this.requestAfter && this.requestAfter > now) {
+      const delayMs = this.requestAfter - now;
+      await new Promise((resolve) => setTimeout(resolve, delayMs));
+    }
+    return super.fetch(request).then((response: any) => {
+      this.requestAfter = this.getRequestAfter(response);
+      return response;
+    });
+  }
+
+  getRateLimitReset(response: any) {
+    return response.headers.get('x-rate-limit-reset');
+  }
+
+  getRateLimitRemaining(response: any) {
+    return response.headers.get('x-rate-limit-remaining');
+  }
+
+  getRequestAfter(response: any) {
+    const rateLimitRemaining = this.getRateLimitRemaining(response);
+    if (rateLimitRemaining <= this.minimumRateLimitRemaining) {
+      const requestAfter =
+        new Date(
+          parseInt(this.getRateLimitReset(response), 10) * 1000,
+        ).getTime() + 1000;
+      this.logger.info(
+        {
+          minimumRateLimitRemaining: this.minimumRateLimitRemaining,
+          requestAfter,
+          url: response.url,
+        },
+        'Minimum `x-rate-limit-remaining` header reached. Temporarily throttling requests',
+      );
+      return requestAfter;
+    }
+    return undefined;
+  }
+}
 
 export default function createOktaClient(
   logger: IntegrationLogger,
   config: OktaIntegrationConfig,
+  options?: RequestExecutorWithEarlyRateLimitingOptions,
 ): OktaClient {
-  const defaultRequestExecutor = new okta.DefaultRequestExecutor({
-    maxRetries: OKTA_MAX_RETRIES,
-    requestTimeout: OKTA_REQUEST_TIMEOUT,
-  });
+  const defaultRequestExecutor = new RequestExecutorWithEarlyRateLimiting(
+    logger,
+    options,
+  );
 
   defaultRequestExecutor.on(
     'backoff',